added e2b emulating test server and unit test for sandbox
